test benchmark naive popcnt

+++ utils.h
@@ -4,7 +4,7 @@
 #include <m4ri.h>
 
 // Shuffle the size first entries of the values array
-void fisher_yates_shuffle(rci_t* values, uint32_t size);
+void fisher_yates_shuffle(rci_t* values, size_t size);
 
 // Load the challenge given by the filename
 // return 1 if everything went fine, 0 otherwise
